feat: allow to unpin dependencies
This comes in handy if one of the configured repositories has pinned all
or most of its dependencies and the others haven't (library vs app).
In case a dependency, which has been pinned by one of the repositories,
publishes a new version to npm and we then execute canarist, the
dependency would be installed in two different versions (the pinned and
the latest version) - which could lead to problems.
With the "unpin" option, we can turn all pinned dependencies into semver
ranges ("^") when running canarist to avoid this problem.